requirements.txt
Django==2.2.5
djangorestframework==3.10.3
drf-yasg==1.16.1
Flask==1.1.1
holdup==1.8.0
psycopg2-binary
pytz==2019.2
sqlparse==0.3.0
packaging==19.2
aiohttp==3.6.1
